+++ src/Command/SyncCommand.php
@@ -77,7 +77,7 @@ class SyncCommand extends Command
             foreach ($projects as $project) {
                 $io->writeln("Processing issues for {$project->getName()}");
 
-                $this->dataSynchronizationService->syncIssuesForProject($project->getId(), function ($i, $length) use ($io) {
+                $this->dataSynchronizationService->syncIssuesForProject($project->getId(), $dataProvider, function ($i, $length) use ($io) {
                     if (0 == $i) {
                         $io->progressStart($length);
                     } elseif ($i >= $length - 1) {
@@ -85,7 +85,7 @@ class SyncCommand extends Command
                     } else {
                         $io->progressAdvance();
                     }
-                }, $dataProvider);
+                });
 
                 $io->writeln('');
             }
@@ -99,7 +99,7 @@ class SyncCommand extends Command
             foreach ($projects as $project) {
                 $io->writeln("Processing worklogs for {$project->getName()}");
 
-                $this->dataSynchronizationService->syncWorklogsForProject($project->getId(), function ($i, $length) use ($io) {
+                $this->dataSynchronizationService->syncWorklogsForProject($project->getId(), $dataProvider, function ($i, $length) use ($io) {
                     if (0 == $i) {
                         $io->progressStart($length);
                     } elseif ($i >= $length - 1) {
@@ -107,7 +107,7 @@ class SyncCommand extends Command
                     } else {
                         $io->progressAdvance();
                     }
-                }, $dataProvider);
+                });
 
                 $io->writeln('');
             }