+++ src/Controller/ProjectController.php
@@ -3,8 +3,6 @@
 namespace App\Controller;
 
 use App\Entity\Project;
-use App\Exception\EconomicsException;
-use App\Exception\UnsupportedDataProviderException;
 use App\Form\ProjectFilterType;
 use App\Form\ProjectType;
 use App\Model\Invoices\ProjectFilterData;
@@ -88,10 +86,6 @@ class ProjectController extends AbstractController
         }
     }
 
-    /**
-     * @throws EconomicsException
-     * @throws UnsupportedDataProviderException
-     */
     #[Route('/{id}/sync', name: 'app_project_sync', methods: ['POST'])]
     public function sync(Project $project, DataSynchronizationService $dataSynchronizationService): Response
     {
@@ -105,8 +99,8 @@ class ProjectController extends AbstractController
             $dataProvider = $project->getDataProvider();
 
             if (null != $dataProvider) {
-                $dataSynchronizationService->syncIssuesForProject($projectId, null, $dataProvider);
-                $dataSynchronizationService->syncWorklogsForProject($projectId, null, $dataProvider);
+                $dataSynchronizationService->syncIssuesForProject($projectId, $dataProvider);
+                $dataSynchronizationService->syncWorklogsForProject($projectId, $dataProvider);
             }
 
             return new JsonResponse([], 200);
@@ -116,5 +110,16 @@ class ProjectController extends AbstractController
                 (int) ($exception->getCode() > 0 ? $exception->getCode() : 500)
             );
         }
+    }
+
+    #[Route('/options', name: 'app_project_options', methods: ['GET'])]
+    public function options(ProjectRepository $projectRepository): JsonResponse
+    {
+        $projects = $projectRepository->getIncluded()->getQuery()->getResult();
+
+        return new JsonResponse(array_map(fn ($project) => [
+            'id' => $project->getId(),
+            'title' => $project->getName(),
+        ], $projects));
     }
 }