+++ src/Controller/SubscriptionController.php
@@ -114,7 +114,7 @@ class SubscriptionController extends AbstractController
                 } else {
                     return new JsonResponse([], 200);
                 }
-                break;
+                // no break
             default:
                 return new JsonResponse(
                     ['error' => 'Unsupported report type'],